回租报价：下拉选api已完成

+++ provider/provider-order-api/src/main/java/com/funtl/myshop/plus/provider/api/ItemCodeService.java
@@ -21,5 +21,5 @@ public interface ItemCodeService{
      * @param itemType
      * @return
      */
-    List<LeasebackSelect> selectItemName(Integer itemType,Integer num);
+    List<LeasebackSelect> selectItemName(Integer itemType,Long num);
 }
